+++ HISTORY.rst
@@ -7,12 +7,13 @@ History
 0.1.44 (2018-04-??)
 ______________________
 
+* Fixed Mouser "quote price" exception in the price tiers.
 * Fixed wxPython exception import.
 * Use the datasheet link information from KiCad and other EDAs, given by 'datasheet' field.
 * Now automatically merge 'description' and other fields to create the groups.
 * GUI save last position and size and others improvements.
 * Display additional information from the web page distributors and use as comment in the ``cat#`` column (just implemented on DigiKey yet).
-* Now is possible to specify country/currency to be priorizede on the distributors scrapes (just implemented on DigiKy yet).
+* Now is possible to specify country/currency to be priorized on the distributors scrapes (just implemented on DigiKey yet).
 
 
 0.1.43 (2018-03-15)